Move all reports used by UI to Cas1ReportService
This helps clarify which reports are CAS1 specific
